# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PathNexis AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SignupBody {
  name: string;
  email: string;
  password: string;
  college?: string;
  year?: string;
  skills?: string[];
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  college?: string | null;
  /** @nullable */
  year?: string | null;
  skills: string[];
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface ResumeAnalyzeBody {
  content: string;
  /** @nullable */
  targetRole?: string | null;
}

export interface ResumeAnalysis {
  id: number;
  score: number;
  suggestions: string[];
  strengths: string[];
  skillGaps: string[];
  /** @nullable */
  targetRole?: string | null;
  createdAt: string;
}

export interface RoadmapGenerateBody {
  targetRole: string;
  currentSkills?: string[];
}

export interface RoadmapStep {
  skill: string;
  description: string;
  resources: string[];
  completed: boolean;
  order: number;
}

export interface Roadmap {
  id: number;
  role: string;
  steps: RoadmapStep[];
  completedCount: number;
  totalCount: number;
  createdAt: string;
}

export interface RoadmapProgressBody {
  roadmapId: number;
  skillIndex: number;
  completed: boolean;
}

export interface Mentor {
  id: number;
  name: string;
  title: string;
  company: string;
  domain: string;
  college: string;
  experience: number;
  bio: string;
  skills: string[];
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  linkedinUrl?: string | null;
  sessionsCount: number;
  rating: number;
}

export interface ConnectionRequest {
  id: number;
  mentorId: number;
  userId: number;
  status: string;
  createdAt: string;
}

export interface Referral {
  id: number;
  title: string;
  company: string;
  domain: string;
  location: string;
  /** @nullable */
  salary?: string | null;
  description: string;
  skills: string[];
  referredBy: string;
  referrerTitle: string;
  postedAt: string;
  /** @nullable */
  deadline?: string | null;
  applicantsCount: number;
}

export interface ReferralApplication {
  id: number;
  referralId: number;
  userId: number;
  status: string;
  appliedAt: string;
  referral: Referral;
}

export interface InterviewStartBody {
  role: string;
  difficulty?: string;
}

export interface InterviewQuestion {
  id: number;
  question: string;
  category: string;
}

export interface InterviewSession {
  sessionId: number;
  role: string;
  difficulty: string;
  questions: InterviewQuestion[];
}

export type InterviewSubmitBodyAnswersItem = {
  questionId: number;
  answer: string;
};

export interface InterviewSubmitBody {
  sessionId: number;
  answers: InterviewSubmitBodyAnswersItem[];
}

export type InterviewFeedbackQuestionFeedbackItem = {
  questionId: number;
  score: number;
  feedback: string;
};

export interface InterviewFeedback {
  sessionId: number;
  overallScore: number;
  feedback: string;
  strengths: string[];
  improvements: string[];
  questionFeedback: InterviewFeedbackQuestionFeedbackItem[];
}

export type DashboardSummaryRecentActivityItem = {
  type: string;
  description: string;
  timestamp: string;
};

export interface DashboardSummary {
  /** @nullable */
  resumeScore?: number | null;
  /** @nullable */
  roadmapProgress?: number | null;
  /** @nullable */
  roadmapRole?: string | null;
  mentorConnections: number;
  jobApplications: number;
  interviewsCompleted: number;
  skillGaps: string[];
  recentActivity: DashboardSummaryRecentActivityItem[];
}

export type ListMentorsParams = {
  domain?: string;
};

export type ListReferralsParams = {
  domain?: string;
};