# lib/api-zod/src/generated/types/interviewFeedback.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PathNexis AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InterviewFeedbackQuestionFeedbackItem } from "./interviewFeedbackQuestionFeedbackItem";

export interface InterviewFeedback {
  sessionId: number;
  overallScore: number;
  feedback: string;
  strengths: string[];
  improvements: string[];
  questionFeedback: InterviewFeedbackQuestionFeedbackItem[];
}