# lib/api-zod/src/generated/types/interviewStartBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PathNexis AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InterviewStartBody {
  role: string;
  difficulty?: string;
}