# lib/api-zod/src/generated/types/mentor.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PathNexis AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Mentor {
  id: number;
  name: string;
  title: string;
  company: string;
  domain: string;
  college: string;
  experience: number;
  bio: string;
  skills: string[];
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  linkedinUrl?: string | null;
  sessionsCount: number;
  rating: number;
}