# lib/api-zod/src/generated/types/referral.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PathNexis AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Referral {
  id: number;
  title: string;
  company: string;
  domain: string;
  location: string;
  /** @nullable */
  salary?: string | null;
  description: string;
  skills: string[];
  referredBy: string;
  referrerTitle: string;
  postedAt: string;
  /** @nullable */
  deadline?: string | null;
  applicantsCount: number;
}