# lib/api-zod/src/generated/types/referralApplication.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PathNexis AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Referral } from "./referral";

export interface ReferralApplication {
  id: number;
  referralId: number;
  userId: number;
  status: string;
  appliedAt: string;
  referral: Referral;
}