# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PathNexis AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  college?: string | null;
  /** @nullable */
  year?: string | null;
  skills: string[];
  createdAt: string;
}